% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_table.R
\name{clean_byGroup}
\alias{clean_byGroup}
\title{clean table by the group}
\usage{
clean_byGroup(meta_table, group)
}
\arguments{
\item{meta_table}{tibble}

\item{group}{i have no idea}
}
\value{
tibble
}
\description{
clean table by the group
}
\examples{
data(combined_table)
clean_table_result <- clean_byGroup(combined_table, treatment)
}
