% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSet.R
\docType{data}
\name{combined_table}
\alias{combined_table}
\title{the ko_abundance table combined with metadata}
\format{
A tibble with 48 rows and 304 variables:
\describe{
  \item{sample}{sample name}
  \item{K00344}{the KO counts number}
  ...
}
}
\source{
\url{http://www.nowhere3.info/}
}
\usage{
combined_table
}
\description{
A dataset containing the ko_abundance and metadata
}
\keyword{datasets}
