% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSet.R
\docType{data}
\name{ko_abundance}
\alias{ko_abundance}
\title{the KEGG table with counts number.}
\format{
A tibble with 300 rows and 49 variables:
\describe{
  \item{ko_id}{the KEGG- KO id}
  \item{R34}{the sample name}
  ...
}
}
\source{
\url{http://www.nowhere2.info/}
}
\usage{
ko_abundance
}
\description{
A dataset containing the sample names and KO counts number
}
\keyword{datasets}
